Make repo work with vagrant
This repo now adds a Vagrantfile that will use this repo to provision a
vagrant box using this repo. Meta.

Main things that needed to change were moving from using group_vars to
role_vars, which is a better fit for this sort of data anyhow.

Also needed to update some tasks to work correctly from within a vm -
stop using the data for home directory because ansible seems to do the
right thing with ~ anyhow.